Add new test for content block
