--- Tests/FeatherWidgets.TestUI.TestCases/ContentBlocks/ContentBlockWidgetOnPage.cs
@@ -1,61 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using ArtOfTest.WebAii.Controls.HtmlControls;
-using ArtOfTest.WebAii.Core;
-using Feather.Widgets.TestUI.Framework;
-using FeatherWidgets.TestUI.TestCases;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Telerik.WebAii.Controls;
-
-namespace FeatherWidgets.TestUI.TestCases.ContentBlocks
-{
-    /// <summary>
-    /// ContentBlockWidgetOnPage test class.
-    /// </summary>
-    [TestClass]
-    public class ContentBlockWidgetOnPage_ : FeatherTestCase
-    {
-        /// <summary>
-        /// UI test ContentBlockWidgetOnPage
-        /// </summary>
-        [TestMethod,
-        Owner(FeatherTeams.FeatherTeam),
-        TestCategory(FeatherTestCategories.ContentBlock)]
-        public void ContentBlockWidgetOnPage()
-        {
-            BAT.Macros().NavigateTo().Pages(this.Culture);
-            BAT.Wrappers().Backend().Pages().PagesWrapper().OpenPageZoneEditor(PageName);
-            BATFeather.Wrappers().Backend().Pages().PageZoneEditorWrapper().AddMvcWidgetHybridModePage(WidgetName);
-            BATFeather.Wrappers().Backend().Pages().PageZoneEditorWrapper().EditWidget(WidgetName);
-            BATFeather.Wrappers().Backend().ContentBlocks().ContentBlocksWrapper().FillContentToContentBlockWidget(ContentBlockContent);
-            BATFeather.Wrappers().Backend().ContentBlocks().ContentBlocksWrapper().SaveChanges();
-            BAT.Wrappers().Backend().Pages().PageZoneEditorWrapper().PublishPage();
-            BAT.Macros().NavigateTo().CustomPage("~/" + PageName.ToLower(), true, this.Culture);
-            BATFeather.Wrappers().Frontend().ContentBlock().ContentBlockWrapper().VerifyContentOfContentBlockOnThePageFrontend(ContentBlockContent);
-        }
-
-        /// <summary>
-        /// Performs Server Setup and prepare the system with needed data.
-        /// </summary>
-        protected override void ServerSetup()
-        {
-            BAT.Macros().User().EnsureAdminLoggedIn();
-            BAT.Arrange(this.TestName).ExecuteSetUp();
-        }
-
-        /// <summary>
-        /// Performs clean up and clears all data created by the test.
-        /// </summary>
-        protected override void ServerCleanup()
-        {
-            BAT.Arrange(this.TestName).ExecuteTearDown();
-        }
-
-        private const string PageName = "ContentBlock";
-        private const string ContentBlockContent = "Test content";
-        private const string WidgetName = "Content block";
-    }
-}
